Add Code of Conduct, adopted from Geek Feminism wiki
Summary: adopts the example CoC document from Geek Feminism without any
substantial changes. Uses the term "Community Team" in all places where
GF's document refers to either the RESPONSE TEAM (individuals empowered
to respond to incidents of hgarassment on bahalf of the group) or the
LEADERSHIP TEAM (which doesn't have any special meaning in this
document, but would refer to the governing body in an organization which
had one). Lists three email addresses as contact points for the Community
Team.

We'll need to follow up with a CONTRIBUTORS document that links to this
Code of Conduct, as well as provides guidance on how to get involved
either in person or online, how to contribute code, and how we would add
people to the project and the Community Team.