Add API endpoint for Twilio to post incoming messages to
Any Twilio payload posted to /api/message will create a new Message post
in the database. If the phone number matches a known user, that user
will be assigned as the author for the post; if not, a new user will be
created.

All messages in the database can be retrieved by a GET request to
/api/messages. (For now this doesn't require authentication to make
debugging easier, but obviously it should be authenticated in
production.)