Add configure script to update Twilio configurations
Adds a script file which will prompt for all the setup values needed and
post the new configuration to Twilio - this is helpful because the ngrok
subdomain changes each time it's run, and otherwise changing the Twilio
configuration through the webn interface is cumbersome and error-prone.

Also updates the INSTALL instructions to cut out unnecessary detail and
explain the steps required a little better.