Hotfix: load all of store in UserProfile component
Previously, loading a user profile would throw a bunch of errors if
state.users and state.tags weren't already populated. This commit fixes
that by injecting the actions for fetching users and tags into the user
profile component. (Will need to update with smarter fetching when the
store is built out more.)