Implement threading of incoming messages
Stores a field called "last_msg" on the User object, and if new incoming
messages come in from a user, they will be considered a response to that
message. Still to do: a way to expire this field (if an event ends, for
example, or if a message comes in that should be handled separately.