Initial (not working yet) socket.io integration
Adds Flask-SocketIO as a websocket server, and redux-socket.io as a
client-side middleware.

This seems to make the connection properly, and as far as I can tell the
message is emitted properly with the gunicorn server used on heroku, but
I can't make it work locally. Leaving this for now to work on something
that I actually know how to do.