Add "event status" field
Because much of the API logic depends on whether an event is "active" or
"archived", this seemed like the next step to add here. Just a basic API
method to change the status of an event, and an indicator on the Event
details page as to whether the event is active or archived.