Add "shutdown" endpoint for development
Because the development server has to run as a multi-threaded app in
order to server the webhooks now, it's really hard to shut it down when
done working with it. This little snippet adds an endpoint which you can
post to and shut it down.

    http post :5000/api/shutdown

Should go without saying that this is just a hack for development and
needs to be removed before this ever goes to production.